Separate folder Prompt function and move to another file
