+++ src/cli.js
@@ -1,8 +1,8 @@
 import arg from 'arg';
 import inquirer from 'inquirer';
-import { createProject } from './main';
 import chalk from 'chalk';
-import fs from 'fs';
+import { folderNamePrompt } from './prompts/folder';
+import { createProject } from './main';
 
 let parseArgumentsIntoOptions = (rawArgs) => {
 
@@ -39,98 +39,8 @@ let parseArgumentsIntoOptions = (rawArgs) => {
   }
 }
 
-let promptForMissingOptions = async (options) => {
-  let [defaultFolderName, defaultTemplate] = ['node-mongo-kit', 'esm'];
-
-  const folderQuestions = [];
-
-  let questionPush = (msgString, folder) => {
-    folderQuestions.push({
-      type: 'input',
-      name: 'folderName',
-      message: msgString,
-      default: folder
-    });
-  }
-
-  const rootDir = process.cwd();
-  const rootDirContent = fs.readdirSync(rootDir, (err, files) => {
-    if (err) {
-      throw err;
-    }
-
-    return files;
-  });
-
-  rootDirContent.push('');
-
-  let matchDefaultValue = rootDirContent.filter(content => {
-    return content.match(defaultFolderName);
-  });
-
-  let folderNameAnswers;
-
-  if (!options.folderName) {
-    if (matchDefaultValue.length >= 1) {
-      defaultFolderName = `${defaultFolderName}-${matchDefaultValue.length}`;
-    }
-    questionPush('Please enter folder name:', defaultFolderName);
-    folderNameAnswers = await inquirer.prompt(folderQuestions);
-  }
-
-  if (options.folderName) {
-    try {
-      fs.accessSync(`./${options.folderName}`, fs.constants.F_OK);
-        console.log( chalk.cyanBright(`Folder name: ${options.folderName} already exists, enter a different folder name instead`) );
-        questionPush( 'Enter different folder name:', null);
-        folderNameAnswers = await inquirer.prompt(folderQuestions);
-    } catch (err) {
-      if (err) {
-        folderNameAnswers = {};
-        folderNameAnswers.folderName = options.folderName;
-       }
-    }
-  }
-
-  try {
-    fs.accessSync(`./${folderNameAnswers.folderName}`, fs.constants.F_OK);
-
-    let equalToAtLeastOneFolder;
-
-    do {
-      equalToAtLeastOneFolder = rootDirContent.some(content => {
-        return content === folderNameAnswers.folderName;
-      });
-
-      if (equalToAtLeastOneFolder === true) {
-        if (folderNameAnswers.folderName !== '') {
-          console.log( chalk.cyanBright(`Folder name: "${folderNameAnswers.folderName}" already exists, enter a different folder name instead`) );
-        } else {
-          console.log( chalk.cyanBright(`Folder name cannot be empty`) );
-        }
-        folderQuestions.push({
-          type: 'input',
-          name: 'folderName',
-          message: 'Enter different folder name:',
-        });
-        if (options.folderName) {
-          folderNameAnswers = await inquirer.prompt(folderQuestions);
-        } else {
-          folderNameAnswers = await inquirer.prompt([folderQuestions[1]]);
-        }
-      }
-    } while (equalToAtLeastOneFolder === true);
-
-  } catch (err) {
-    if (err) {
-      //Dummy if statement to prevent: unhandledPromiseRejectionWarning in console
-    }
-  }
-
-  //Note: This affects only the try block of the previous if (options.folderName) statement
-  if (options.folderName) {
-    options.folderName = folderNameAnswers.folderName;
-  }
+let promptForMissingOptions = async (options, folderNameAnswers) => {
+  let defaultTemplate = 'esm';
 
   const templateQuestions = [];
 
@@ -204,7 +114,9 @@ export let cli = async (args) => {
     options.git = false;
   }
 
-  options = await promptForMissingOptions(options);
+  let [updatedOptions, folderNameAnswers] = await folderNamePrompt(options);
+
+  options = await promptForMissingOptions(updatedOptions, folderNameAnswers);
 
   //console.log(options);
